requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
jinja2==3.1.6
sqlalchemy==2.0.48
pydantic-settings==2.13.1
python-multipart==0.0.22
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
email-validator==2.3.0
httpx==0.28.1
markdown==3.10.2
bleach==6.3.0
reportlab==4.4.10
python-slugify==8.0.4
authlib==1.6.9
pytest==9.0.2
itsdangerous==2.2.0
